+++ src/harnest/store.py
@@ -3,5 +3,6 @@
 from .checkpoint import MemoryStore
 from .store_postgres import PostgresStore
 from .store_redis import RedisStore
+from .storage_registry import CustomStorage, StorageRegistry
 
-__all__ = ["MemoryStore", "PostgresStore", "RedisStore"]
+__all__ = ["MemoryStore", "PostgresStore", "RedisStore", "CustomStorage", "StorageRegistry"]